Moved App.vue into pages as MainPage
