+++ src/pages/MainPage.vue
@@ -191,17 +191,17 @@
 
 
 <script>
-import Audit from './components/Audit.vue'
-import ClassSearch from './components/ClassSearch.vue'
-import Road from './components/Road.vue'
-import FilterSet from "./components/FilterSet.vue"
-import RoadTabs from "./components/RoadTabs.vue"
-import ConflictDialog from "./components/ConflictDialog.vue"
-import Auth from "./components/Auth.vue"
+import Audit from './../components/Audit.vue'
+import ClassSearch from './../components/ClassSearch.vue'
+import Road from './../components/Road.vue'
+import FilterSet from "./../components/FilterSet.vue"
+import RoadTabs from "./../components/RoadTabs.vue"
+import ConflictDialog from "./../components/ConflictDialog.vue"
+import Auth from "./../components/Auth.vue"
 import $ from 'jquery'
 import Vue from 'vue'
-import ClassInfo from "./components/ClassInfo.vue"
-import ImportExport from "./components/ImportExport.vue"
+import ClassInfo from "./../components/ClassInfo.vue"
+import ImportExport from "./../components/ImportExport.vue"
 
 var MAIN_URL = "http://localhost:8080"
 var DATE_FORMAT = "YYYY-MM-DDTHH:mm:ss.SSS000Z"
